Unfold TestLocaleClass (and other Locale class-related tests) to test_locale_class
